% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADComorbidity.R
\name{DADComorbidity}
\alias{DADComorbidity}
\title{DADComorbidity returns the comorbidities for a given disease.}
\usage{
DADComorbidity(disease, spark_data_frame = NULL)
}
\arguments{
\item{spark_data_frame}{}
}
\description{
DADComorbidity returns the comorbidities for a given disease.
}
