% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADload.R
\name{DADload}
\alias{DADload}
\title{DADload: Loads the DAD data file and saves as an RDA file}
\usage{
DADload(savfile, rdafile = "")
}
\arguments{
\item{savfile}{The full path to the DAD SPSS file}

\item{rdafile}{The full path to the Rda file created. dadR in working dir if empty.}
}
\description{
DADload: Loads the DAD data file and saves as an RDA file
}
\examples{
DADload("~/clin_sample_spss.sav")
}
