% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADSameDisease.R
\name{DADSameDisease}
\alias{DADSameDisease}
\title{Title DADSameDisease - List all diseases with given diagnosis.}
\usage{
DADSameDisease(disease, spark_data_frame = NULL)
}
\arguments{
\item{spark_data_frame}{}
}
\value{
spark_data_frame with records
}
\description{
Title DADSameDisease - List all diseases with given diagnosis.
}
\examples{
spark_df <- DADSameDisease("J08")
  r_df <- collect(spark_df)
  r_dt <- as.data.table(r_df)
  r_dt  # All records with the diagnosis J08
}
