% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADSameTreatment.R
\name{DADSameTreatment}
\alias{DADSameTreatment}
\title{Title DADSametreatment - List all records with given intervention.}
\usage{
DADSameTreatment(treatment, spark_data_frame = NULL)
}
\arguments{
\item{spark_data_frame}{}
}
\value{
spark_data_frame with records
}
\description{
Title DADSametreatment - List all records with given intervention.
}
\examples{
spark_df <- DADSameTreatment(" 1QE35JAG7")
  r_df <- collect(spark_df)
  r_dt <- as.data.table(r_df)
  r_dt  # All records with the intervention 1QE35JAG7
}
